requirements.txt
opencv-python
mediapipe
ultralytics
google-cloud-vision
python-dotenv
sqlalchemy
psycopg2-binary